refactor: split the Studio BFF god-module and App state
Break the ~1900-line server.py into focused modules and consolidate the
front's per-session state. Behaviour preserved.

- studio BFF: extract paths.py (jail + workspace/history), registry.py
  (a SessionState dataclass held in one dict[str, SessionState], replacing
  the 8 parallel dicts; sessions.json schema unchanged), auth.py, terminal.py,
  tensorboard.py, jobs.py. server.py (866 lines) is now app creation + route
  wiring; entangled experiment/job-control routes stay there by design.
- studio front: consolidate nine per-session Record maps into one
  Record<string, SessionUiState> (sessionState.ts) with patchSession /
  replaceSessionField helpers.

Verified: every route group serves 200 on relaunch, the registry round-trips
the legacy sessions.json, the smoke tests pass, and tsc + build are clean.
